+++ src/Attestation/Attestation/Models/AttestationManagementClient.cs
@@ -13,23 +13,25 @@
 // ----------------------------------------------------------------------------------
 
 using System;
+using System.Collections.Generic;
 using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Management.Attestation;
 using Microsoft.Azure.Management.Attestation.Models;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
-
+using System.Security.Cryptography.X509Certificates;
+using Microsoft.IdentityModel.JsonWebTokens;
 namespace Microsoft.Azure.Commands.Attestation.Models
 {
-    public class AttestationClient
+    public class AttestationManagementClient
     {
-        private readonly AttestationManagementClient attestationClient;
+        private readonly Management.Attestation.AttestationManagementClient attestationClient;
 
-        public AttestationClient(IAzureContext context)
+        public AttestationManagementClient(IAzureContext context)
         {
-            attestationClient = AzureSession.Instance.ClientFactory.CreateArmClient<AttestationManagementClient>(context, AzureEnvironment.Endpoint.ResourceManager);
+            attestationClient = AzureSession.Instance.ClientFactory.CreateArmClient<Management.Attestation.AttestationManagementClient>(context, AzureEnvironment.Endpoint.ResourceManager);
         }
 
-        public AttestationClient()
+        public AttestationManagementClient()
         {
         }
         public PSAttestation CreateNewAttestation(AttestationCreationParameters parameters)
@@ -38,11 +40,11 @@ namespace Microsoft.Azure.Commands.Attestation.Models
             {
                 throw new ArgumentNullException("parameters");
             }
-            if (string.IsNullOrWhiteSpace(parameters.ProviderName))
+            if (string.IsNullOrEmpty(parameters.ProviderName))
             {
                 throw new ArgumentNullException("parameters.ProviderName");
             }
-            if (string.IsNullOrWhiteSpace(parameters.ResourceGroupName))
+            if (string.IsNullOrEmpty(parameters.ResourceGroupName))
             {
                 throw new ArgumentNullException("parameters.ResourceGroupName");
             }
@@ -51,6 +53,11 @@ namespace Microsoft.Azure.Commands.Attestation.Models
             {
                 _creationParams.AttestationPolicy = parameters.AttestationPolicy;
             }
+            if (parameters.PolicySigningCertificates != null)
+            {
+                _creationParams.PolicySigningCertificates = parameters.PolicySigningCertificates;
+            }
+
 
             var response = attestationClient.AttestationProviders.Create(
                 resourceGroupName: parameters.ResourceGroupName,
@@ -61,11 +68,11 @@ namespace Microsoft.Azure.Commands.Attestation.Models
 
         public PSAttestation GetAttestation(string attestationName, string resourceGroupName)
         {
-            if (string.IsNullOrWhiteSpace(attestationName))
+            if (string.IsNullOrEmpty(attestationName))
             {
                 throw new ArgumentNullException("attestationName");
             }
-            if (string.IsNullOrWhiteSpace(resourceGroupName))
+            if (string.IsNullOrEmpty(resourceGroupName))
             {
                 throw new ArgumentNullException("resourceGroupName");
             }
@@ -74,16 +81,55 @@ namespace Microsoft.Azure.Commands.Attestation.Models
         }
         public void DeleteAttestation(string attestationName, string resourceGroupName)
         {
-            if (string.IsNullOrWhiteSpace(attestationName))
+            if (string.IsNullOrEmpty(attestationName))
             {
                 throw new ArgumentNullException("attestationName");
             }
 
-            if (string.IsNullOrWhiteSpace(resourceGroupName))
+            if (string.IsNullOrEmpty(resourceGroupName))
             {
                 throw new ArgumentNullException("resourceGroupName");
             }
             attestationClient.AttestationProviders.Delete(resourceGroupName, attestationName);
+        }
+
+        public X509Certificate2Collection GetX509CertificateFromPEM(string pemString, string section)
+        {
+            X509Certificate2Collection certificateCollection = new X509Certificate2Collection();
+            var header = String.Format("-----BEGIN {0}-----", section);
+            var footer = String.Format("-----END {0}-----", section);
+
+            var start = 0;
+            var lengthOfSection = 0;
+            while (true)
+            {
+                start = pemString.IndexOf(header, StringComparison.Ordinal);
+
+                if (start < 0)
+                    break;
+                start += header.Length;
+                lengthOfSection = pemString.IndexOf(footer, start, StringComparison.Ordinal) - start;
+                if (lengthOfSection < 0)
+                    break;
+                byte [] certBuffer = Convert.FromBase64String(pemString.Substring(start, lengthOfSection));
+                X509Certificate2 certs = new X509Certificate2(certBuffer);
+                certificateCollection.Add(certs);
+                pemString = pemString.Substring(start + lengthOfSection);
+            }
+            return certificateCollection;
+        }
+
+        public JSONWebKeySet GetJSONWebKeySet(X509Certificate2Collection certificateCollection)
+        {
+            var jwks = new JSONWebKeySet();
+            jwks.Keys = new List<JSONWebKey>();
+            foreach (var certificate in certificateCollection)
+            {
+                var jwk = new JSONWebKey() { Kty = "RSA" };
+                jwk.X5c = new List<string>() { System.Convert.ToBase64String(certificate.Export(X509ContentType.Cert)) };
+                jwks.Keys.Add(jwk);
+            }
+            return jwks;
         }
     }
 }