+++ src/Network/Network/ChangeLog.md
@@ -20,6 +20,9 @@
 
 ## Upcoming Release
 * Change `Start-AzVirtualNetworkGatewayConnectionPacketCapture.md` and `Start-AzVirtualnetworkGatewayPacketCapture.md` FilterData option examples.
+* Add `PrivateRange` parameter to `AzureFirewall`
+	- Updated cmdlet:
+		- New-AzFirewall
 
 ## Version 2.0.0
 * Change all cmdlets for PrivateEndpointConnection to support generic service provider.