+++ src/Network/Network/ChangeLog.md
@@ -20,10 +20,10 @@
 
 ## Upcoming Release
 * Update references in .psd1 to use relative path
+* Support for IpGroups in AzureFirewall Application,Nat & Network Rules.
 
 ## Version 2.1.0
 * Change `Start-AzVirtualNetworkGatewayConnectionPacketCapture.md` and `Start-AzVirtualnetworkGatewayPacketCapture.md` FilterData option examples.
-* Support for IpGroups in AzureFirewall Application,Nat & Network Rules.
 * Add `PrivateRange` parameter to `AzureFirewall`
 	- Updated cmdlet:
 		- New-AzFirewall